Restructure Serial/MappingNearestProjection
