+++ tests/serial/mapping-nearest-projection/helpers.cpp
@@ -1,0 +1,109 @@
+#ifndef PRECICE_NO_MPI
+
+#include "helpers.hpp"
+#include "testing/Testing.hpp"
+
+#include "precice/SolverInterface.hpp"
+
+void testMappingNearestProjection(bool defineEdgesExplicitly, const std::string configFile, const TestContext &context)
+{
+  using Eigen::Vector3d;
+
+  const double z = 0.3;
+
+  // MeshOne
+  Vector3d coordOneA{0.0, 0.0, z};
+  Vector3d coordOneB{1.0, 0.0, z};
+  Vector3d coordOneC{1.0, 1.0, z};
+  Vector3d coordOneD{0.0, 1.0, z};
+  double   valOneA = 1.0;
+  double   valOneB = 3.0;
+  double   valOneC = 5.0;
+  double   valOneD = 7.0;
+
+  // MeshTwo
+  Vector3d coordTwoA{0.0, 0.0, z + 0.1};               // Maps to vertex A
+  Vector3d coordTwoB{0.0, 0.5, z - 0.01};              // Maps to edge AD
+  Vector3d coordTwoC{2.0 / 3.0, 1.0 / 3.0, z + 0.001}; // Maps to triangle ABC
+  // This corresponds to the point C from mesh two on the triangle ABC on mesh one.
+  Vector3d barycenterABC{0.3798734633239789, 0.24025307335204216, 0.3798734633239789};
+  double   expectedValTwoA = 1.0;
+  double   expectedValTwoB = 4.0;
+  double   expectedValTwoC = Vector3d{valOneA, valOneB, valOneC}.dot(barycenterABC);
+
+  if (context.isNamed("SolverOne")) {
+    precice::SolverInterface cplInterface("SolverOne", configFile, 0, 1);
+    // namespace is required because we are outside the fixture
+    const int meshOneID = cplInterface.getMeshID("MeshOne");
+
+    // Setup mesh one.
+    int idA = cplInterface.setMeshVertex(meshOneID, coordOneA.data());
+    int idB = cplInterface.setMeshVertex(meshOneID, coordOneB.data());
+    int idC = cplInterface.setMeshVertex(meshOneID, coordOneC.data());
+    int idD = cplInterface.setMeshVertex(meshOneID, coordOneD.data());
+
+    if (defineEdgesExplicitly) {
+
+      int idAB = cplInterface.setMeshEdge(meshOneID, idA, idB);
+      int idBC = cplInterface.setMeshEdge(meshOneID, idB, idC);
+      int idCD = cplInterface.setMeshEdge(meshOneID, idC, idD);
+      int idDA = cplInterface.setMeshEdge(meshOneID, idD, idA);
+      int idCA = cplInterface.setMeshEdge(meshOneID, idC, idA);
+
+      cplInterface.setMeshTriangle(meshOneID, idAB, idBC, idCA);
+      cplInterface.setMeshTriangle(meshOneID, idCD, idDA, idCA);
+
+    } else {
+      cplInterface.setMeshTriangleWithEdges(meshOneID, idA, idB, idC);
+      cplInterface.setMeshTriangleWithEdges(meshOneID, idC, idD, idA);
+    }
+
+    // Initialize, thus sending the mesh.
+    double maxDt = cplInterface.initialize();
+    BOOST_TEST(cplInterface.isCouplingOngoing(), "Sending participant should have to advance once!");
+
+    // Write the data to be send.
+    int dataAID = cplInterface.getDataID("DataOne", meshOneID);
+    cplInterface.writeScalarData(dataAID, idA, valOneA);
+    cplInterface.writeScalarData(dataAID, idB, valOneB);
+    cplInterface.writeScalarData(dataAID, idC, valOneC);
+    cplInterface.writeScalarData(dataAID, idD, valOneD);
+
+    // Advance, thus send the data to the receiving partner.
+    cplInterface.advance(maxDt);
+    BOOST_TEST(!cplInterface.isCouplingOngoing(), "Sending participant should have to advance once!");
+    cplInterface.finalize();
+  } else {
+    BOOST_TEST(context.isNamed("SolverTwo"));
+    SolverInterface cplInterface("SolverTwo", configFile, 0, 1);
+    // namespace is required because we are outside the fixture
+    int meshTwoID = cplInterface.getMeshID("MeshTwo");
+
+    // Setup receiving mesh.
+    int idA = cplInterface.setMeshVertex(meshTwoID, coordTwoA.data());
+    int idB = cplInterface.setMeshVertex(meshTwoID, coordTwoB.data());
+    int idC = cplInterface.setMeshVertex(meshTwoID, coordTwoC.data());
+
+    // Initialize, thus receive the data and map.
+    double maxDt = cplInterface.initialize();
+    BOOST_TEST(cplInterface.isCouplingOngoing(), "Receiving participant should have to advance once!");
+
+    // Read the mapped data from the mesh.
+    int    dataAID = cplInterface.getDataID("DataOne", meshTwoID);
+    double valueA, valueB, valueC;
+    cplInterface.readScalarData(dataAID, idA, valueA);
+    cplInterface.readScalarData(dataAID, idB, valueB);
+    cplInterface.readScalarData(dataAID, idC, valueC);
+
+    BOOST_TEST(valueA == expectedValTwoA);
+    BOOST_TEST(valueB == expectedValTwoB);
+    BOOST_TEST(valueC == expectedValTwoC);
+
+    // Verify that there is only one time step necessary.
+    cplInterface.advance(maxDt);
+    BOOST_TEST(!cplInterface.isCouplingOngoing(), "Receiving participant should have to advance once!");
+    cplInterface.finalize();
+  }
+}
+
+#endif
