+++ tests/serial/mapping-nearest-projection/helpers.hpp
@@ -1,0 +1,12 @@
+#ifndef PRECICE_NO_MPI
+
+#pragma once
+
+#include "testing/TestContext.hpp"
+
+using namespace precice;
+using precice::testing::TestContext;
+
+void testMappingNearestProjection(bool defineEdgesExplicitly, const std::string configFile, const TestContext &context);
+
+#endif
